Refactor: RoomBox -> RoomCard 변경 및 구조 개선
1. RoomCard로 컴포넌트 이름을 변경하였습니다.
2. room에 위치시켰습니다.
