--- client/src/components/RoomBox.tsx
@@ -1,164 +1,0 @@
-import styled, { css } from 'styled-components';
-import { RoomInfo } from './main/RoomList';
-import { getRoomByUuid, postEnterRoom } from '@src/apis';
-import { useHistory } from 'react-router';
-import { useCallback, useEffect, useRef } from 'react';
-import socket from '@socket/socket';
-import { SocketEvents } from '@socket/socketEvents';
-import { useUser } from '@contexts/userContext';
-import { RoomType, TOAST_MESSAGE } from '@utils/constant';
-import { ROOM_BOX } from '@utils/styleConstant';
-import { useToast } from '@hooks/useToast';
-
-const RoomBoxWrapper = styled.div`
-  ${({ theme }) => theme.flexCenter};
-  flex-direction: column;
-  background-color: ${({ theme }) => theme.colors.white};
-  padding: ${({ theme }) => theme.paddings.base};
-  width: 100%;
-  height: ${ROOM_BOX.height}rem;
-  border-radius: ${({ theme }) => theme.borderRadius.base};
-  cursor: pointer;
-  position: relative;
-  box-shadow: rgba(0, 0, 0, 0.16) 0px 3px 6px, rgba(0, 0, 0, 0.23) 0px 3px 6px;
-  -webkit-transition: all 0.2s cubic-bezier(0.165, 0.84, 0.44, 1);
-  transition: all 0.2s cubic-bezier(0.165, 0.84, 0.44, 1);
-  &:hover {
-    transform: scale(1.02);
-  }
-  &:active {
-    background: ${({ theme }) => theme.colors.blue};
-    transform: scale(0.98);
-  }
-  &::after {
-    content: '';
-    border-radius: 5px;
-    position: absolute;
-    z-index: -1;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
-    opacity: 0;
-    -webkit-transition: all 0.1s cubic-bezier(0.165, 0.84, 0.44, 1);
-    transition: all 0.1s cubic-bezier(0.165, 0.84, 0.44, 1);
-  }
-  &:hover::after {
-    opacity: 1;
-  }
-`;
-
-const RoomBoxTop = styled.div`
-  width: 100%;
-  height: 60%;
-  margin-top: ${({ theme }) => theme.margins.base};
-  display: flex;
-  flex-direction: column;
-`;
-
-const RoomTopMenu = styled.div`
-  display: flex;
-`;
-
-const RoomTitle = styled.h6`
-  display: block;
-  width: 100%;
-  ${({ theme }) => css`
-    font-size: ${theme.fontSizes.xl};
-    font-weight: ${theme.fontWeights.normal};
-    margin-bottom: ${theme.margins.base};
-  `}
-`;
-
-const RoomDescription = styled.p`
-  font-size: ${({ theme }) => theme.fontSizes.base};
-  opacity: 0.9;
-`;
-
-const RoomBoxBottom = styled.div`
-  width: 100%;
-  height: 40%;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-`;
-
-const RoomFieldType = styled.div`
-  width: 6rem;
-  height: 2rem;
-  ${({ theme }) => css`
-    ${theme.flexCenter};
-    font-size: ${theme.fontSizes.sm};
-    background-color: ${theme.colors.blue};
-    color: ${theme.colors.white};
-    padding: ${theme.paddings.sm};
-    border-radius: ${theme.borderRadius.sm};
-  `}
-`;
-
-const RoomOwnerNickname = styled.span`
-  color: ${({ theme }) => theme.colors.green};
-`;
-
-const RoomAdmitNumber = styled.span``;
-
-interface RoomBoxProps {
-  roomInfo: RoomInfo;
-}
-
-const RoomBox = ({ roomInfo }: RoomBoxProps): JSX.Element => {
-  const { uuid, title, description, nowHeadcount, maxHeadcount, roomType, owner } = roomInfo;
-  const { login, nickname } = useUser();
-  const toast = useToast();
-  const roomDataRef = useRef<RoomInfo>();
-  const history = useHistory();
-
-  const verifyBySocket = useCallback(async () => {
-    socket.emit(SocketEvents.canIEnter, { uuid, nickname });
-  }, [uuid, nickname]);
-
-  const onClickRoomBox = useCallback(async () => {
-    if (!login) {
-      toast('error', TOAST_MESSAGE.notAllowedNonLogin);
-      return;
-    }
-    const { isOk, data } = await getRoomByUuid(uuid);
-    if (isOk && data) {
-      if (data.nowHeadcount >= data.maxHeadcount) return;
-      roomDataRef.current = data;
-      verifyBySocket();
-    }
-  }, [uuid, login, toast, verifyBySocket]);
-
-  const enterRoom = useCallback(() => {
-    const pathname = roomType === RoomType.tadak ? `/room/tadak/${uuid}` : `/room/campfire/${uuid}`;
-    postEnterRoom(uuid);
-    history.push({ pathname, state: roomDataRef.current });
-  }, [history, uuid, roomType, roomDataRef]);
-
-  useEffect(() => {
-    socket.removeListener(SocketEvents.youCanEnter);
-    socket.on(SocketEvents.youCanEnter, enterRoom);
-  }, [enterRoom]);
-
-  return (
-    <RoomBoxWrapper onClick={onClickRoomBox}>
-      <RoomBoxTop>
-        <RoomTopMenu>
-          <RoomTitle>{title}</RoomTitle>
-          <RoomFieldType>{owner?.devField?.name || 'All'}</RoomFieldType>
-        </RoomTopMenu>
-        <RoomDescription>{description}</RoomDescription>
-      </RoomBoxTop>
-      <RoomBoxBottom>
-        <RoomOwnerNickname>호스트 : {owner?.nickname}</RoomOwnerNickname>
-        <RoomAdmitNumber>
-          {nowHeadcount} / {maxHeadcount}
-        </RoomAdmitNumber>
-      </RoomBoxBottom>
-    </RoomBoxWrapper>
-  );
-};
-
-export default RoomBox;
